8182299: Enable disabled clang warnings, build on OSX 10 + Xcode 8 8182656: Make the required changes in GC code to build on OSX 10 + Xcode 8 8182657: Make the required changes in Runtime code to build on OSX 10 + Xcode 8 8182658: Make the required changes in Compiler code to build on OSX 10 + Xcode 8

+++ hotspot/src/share/vm/logging/logFileOutput.cpp
@@ -191,7 +191,7 @@ bool LogFileOutput::parse_options(const char* options, outputStream* errstream) 
       success = Arguments::atojulong(value_str, &value);
       if (!success || (value > SIZE_MAX)) {
         errstream->print_cr("Invalid option: %s must be in range [0, "
-                            SIZE_FORMAT "]", FileSizeOptionKey, SIZE_MAX);
+                            SIZE_FORMAT "]", FileSizeOptionKey, (size_t)SIZE_MAX);
         success = false;
         break;
       }
@@ -436,4 +436,3 @@ void LogFileOutput::describe(outputStream *out) {
              byte_size_in_proper_unit(_rotate_size),
              proper_unit_for_byte_size(_rotate_size));
 }
-
